+++ src/components/badge/badge.tsx
@@ -261,6 +261,7 @@ export const EuiBadge: FunctionComponent<EuiBadgeProps> = ({
             size="s"
             {...closeButtonProps}
             className={closeClassNames}
+            color="default" // forces app icons to inherit de parent color
           />
         </button>
       );
@@ -270,6 +271,7 @@ export const EuiBadge: FunctionComponent<EuiBadgeProps> = ({
           type={iconType}
           size={children ? 's' : 'm'}
           className="euiBadge__icon"
+          color="default" // forces app icons to inherit de parent color
         />
       );
     }