+++ src/components/badge/beta_badge/beta_badge.tsx
@@ -103,6 +103,7 @@ export const EuiBetaBadge: FunctionComponent<EuiBetaBadgeProps> = ({
         type={iconType}
         size="m"
         aria-hidden="true"
+        color="default" // forces app icons to inherit de parent color
       />
     );
   }