+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Bug fixes**
 
 - Fixed missing i18n tokens for `EuiFilePicker` ([#4750](https://github.com/elastic/eui/pull/4750))
+- Fixed a bug in `EuiIcon` where app icons were getting two-tone colors when nested in specific components ([#4750](https://github.com/elastic/eui/pull/4750))
 
 ## [`32.2.0`](https://github.com/elastic/eui/tree/v32.2.0)